requirements.txt
rvc3python
machinevision-toolbox-python
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.0
tqdm>=4.62.0 